fix: constants file merged, fix in imports and texts
